% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSynthesis.R
\name{ABSynthesis}
\alias{ABSynthesis}
\title{Combine AB test results for a summary uplift and p-value}
\usage{
ABSynthesis(successes_base, base_n, successes_variant, variant_n, key = "")
}
\arguments{
\item{successes_base}{Count of responses in the base condition}

\item{base_n}{Count of visitors exposed to the base condition}

\item{successes_variant}{Count of responses in the base condition}

\item{variant_n}{Count of visitors exposed to the base condition}
}
\value{
Summary uplift and p-value of a meta-analysis of online controlled experiments
}
\description{
ABSynthesis combines experiments with binomial outcomes (successes and trials), reporting the summary uplift and p-value.
}
\examples{
ABSynthesis(c(10, 11, 12), c(50, 50, 50), c(12, 13, 14), c(50, 50, 50))
}
