% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingBias.R
\name{MissingBias}
\alias{MissingBias}
\title{Detect whether there is a bias in the missingness of values in a vector.}
\usage{
MissingBias(biased, predictor, email = "", key = "")
}
\arguments{
\item{biased}{Count of responses in the base condition}

\item{predictor}{Count of visitors exposed to the base condition}

\item{email}{The user email provided to receive the API key.}

\item{key}{The API key received from scientificmicroservices.com . Can be left blank if the value is stored as 'scimicro_key' in .Renviron}
}
\value{
A numeric 1 or 0 output indicating whether data is more likely to be missing in the biased column with particular values of the predictor column.
}
\description{
MissingBias returns a 1 (is biased) or 0 (is not biased) indicating whether the values in the biased column are more likely to be missing depending on the value of the predictor column.
}
\examples{
\dontrun{MissingBias(biased = c(1, 4, 6, NA, 10), predictor = c(5, 6, 7, 25, 5), email = 'info@scientificmicroservices.com')}
}
