% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingGraph.R
\name{MissingGraph}
\alias{MissingGraph}
\title{Get a graph of missing values in a dataframe by row and column}
\usage{
MissingGraph(missing_data, email = "", key = "")
}
\arguments{
\item{missing_data}{A data.frame with potentially missing data.}

\item{email}{The user email provided to receive the API key.}

\item{key}{The API key received from scientificmicroservices.com . Can be left blank if the value is stored as 'scimicro_key' in .Renviron}
}
\value{
A summary of percentage of missing data by row and column.
}
\description{
MissingGraph takes a dataframe and returns a png representing the row and column of missing data.
}
\examples{
\dontrun{MissingGraph(missing_data = data.frame(c(10, 11, 12), c(50, NA, 50), c(NA, NA, 14), c(50, 50, NA)), email = 'info@scientificmicroservices.com')}
}
