% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingRowsCols.R
\name{MissingRowsCols}
\alias{MissingRowsCols}
\title{Get a report of missing values in a dataframe by row and column}
\usage{
MissingRowsCols(missing_data, key = "")
}
\arguments{
\item{missing_data}{A data.frame with potentially missing data.}
}
\value{
A summary of percentage of missing data by row and column.
}
\description{
MissingRowsCols takes a dataframe and reports the percentage of missing data by row and column.
}
\examples{
MissingRowsCols(data.frame(c(10, 11, 12), c(50, NA, 50), c(NA, NA, 14), c(50, 50, NA)))
}
